youtube_transcript/youtube-transcript-qa/requirements.txt
streamlit
boto3
openai
python-dotenv
youtube-transcript-api
flotorch_core
